Intents as part of the communications data
This commit intends to allow the GUI to inform the backend what action it is intending to perform by adding a new field: 'intent' to the JSON data. By providing this field, the backend bridge knows what the message sent is intended to do and can quickly call the appropriate channels to perform the request.